+++ packages/agents/src/observability/ai/options.ts
