+++ packages/agents/src/tests/observability/ai-sdk-v7-telemetry.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
-import { RecordingTracer } from "../../../test-support/recording-tracer";
-import { createAISDKV7Telemetry } from "./telemetry";
+import { RecordingTracer } from "./recording-tracer";
+import { createAISDKV7Telemetry } from "../../observability/ai/v7/telemetry";
 
 describe("createAISDKV7Telemetry", () => {
   it("traces operation and model callbacks with call id correlation", () => {
