--- packages/ai-tracing/scripts/build.ts
@@ -1,28 +1,0 @@
-import { build } from "tsdown";
-import { formatDeclarationFiles } from "../../../scripts/format-declarations";
-
-async function main() {
-  await build({
-    clean: true,
-    dts: true,
-    entry: ["src/ai-sdk.ts"],
-    deps: {
-      skipNodeModulesBundle: true,
-      neverBundle: ["cloudflare:workers"]
-    },
-    format: "esm",
-    sourcemap: true,
-    fixedExtension: false
-  });
-
-  // then run oxfmt on the generated .d.ts files
-  formatDeclarationFiles();
-
-  process.exit(0);
-}
-
-main().catch((err) => {
-  // Build failures should fail
-  console.error(err);
-  process.exit(1);
-});
