--- packages/ai-tracing/src/tracing/cloudflare.ts
@@ -1,5 +1,0 @@
-import { tracing as cloudflareTracing } from "cloudflare:workers";
-import { createTracer } from "./tracer";
-import type { Tracer } from "./tracer";
-
-export const tracer: Tracer = createTracer(cloudflareTracing);
